further opt

+++ small/small_LinearAlgebra.c
@@ -13,7 +13,7 @@
 /*        R. Sander, Max-Planck Institute for Chemistry, Mainz, Germany */
 /*                                                                  */
 /* File                 : small_LinearAlgebra.c                     */
-/* Time                 : Sat Jan 29 17:28:18 2022                  */
+/* Time                 : Sun Jan 30 10:01:38 2022                  */
 /* Working directory    : /home/kyriacos/CyprusInstitute/kpp/small  */
 /* Equation file        : small.kpp                                 */
 /* Output root filename : small                                     */
@@ -371,7 +371,7 @@ double WLAMCH( char C )
 	  if (Suma <= ONE) break;
         } /* end for */
         if (i==80) {
-	   printf("\nERROR IN WLAMCH. Very small EPS = %g\n",Eps);
+	   printf("\nERROR IN WLAMCH. Very small EPS = %g\n",(double)Eps);
            return (double)2.2e-16;
 	}
         Eps *= TWO; i--;
